requirements.txt
fastapi
uvicorn[standard]
boto3
botocore
langchain
openai
pydantic
langchain-community
python-dotenv
langchain-core
langgraph
